+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Properties related to IDP, SAML and OIDC are deprecated in `konnect_portal_auth` resource
 * Properties `config.tools[*].headers.key` and `config.tools[*].query.key` in `konnect_gateway_plugin_ai_mcp_proxy` resource now accept list of strings
 * Property `config.limits.key` in `konnect_gateway_plugin_response_ratelimiting` resource now accepts an object
+* Fix panic when dynamic ordering is used with custom plugin
 
 ## 3.4.2
 > Released on 2025/11/11